+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/assets/infrastructure/adapters/rest/metrics/registrylookup/RegistryLookupMetricsController.java
@@ -22,10 +22,15 @@
 package org.eclipse.tractusx.traceability.assets.infrastructure.adapters.rest.metrics.registrylookup;
 
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.ArraySchema;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.validation.constraints.Size;
+import org.eclipse.tractusx.traceability.assets.domain.model.Asset;
 import org.eclipse.tractusx.traceability.assets.infrastructure.adapters.metrics.RegistryLookupMeterRegistry;
 import org.eclipse.tractusx.traceability.assets.infrastructure.adapters.metrics.RegistryLookupMetric;
 import org.eclipse.tractusx.traceability.common.model.PageResult;
@@ -52,11 +57,14 @@ public class RegistryLookupMetricsController {
 		tags = {"Registry"},
 		description = "The endpoint gets metrics for database.",
 		security = @SecurityRequirement(name = "oAuth2", scopes = "profile email"))
-	@ApiResponses(value = {@ApiResponse(responseCode = "200", description = "OK."),
+    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "Returns the paged result found for Asset", content = @Content(
+            mediaType = "application/json",
+            array = @ArraySchema(arraySchema = @Schema(description = "RegistryLookupMetric", implementation = RegistryLookupMetric.class), maxItems = Integer.MAX_VALUE)
+    )),
 		@ApiResponse(responseCode = "401", description = "Authorization failed."),
 		@ApiResponse(responseCode = "403", description = "Forbidden.")})
 	@GetMapping("/registry-lookup")
-	public PageResult<RegistryLookupMetric> metrics(Pageable pageable) {
+	public @Size(max = 1000) @ArraySchema(arraySchema = @Schema(description = "RegistryLookupMetric", implementation = RegistryLookupMetric.class), maxItems = Integer.MAX_VALUE) PageResult<RegistryLookupMetric> metrics(@Size(max = 1000) Pageable pageable) {
 		return registryLookupMeterRegistry.getMetrics(pageable);
 	}
 }