+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/bpn/mapping/domain/model/BpnEdcMapping.java
@@ -20,7 +20,10 @@
 package org.eclipse.tractusx.traceability.bpn.mapping.domain.model;
 
 import io.swagger.annotations.ApiModelProperty;
+import io.swagger.v3.oas.annotations.media.ArraySchema;
+import io.swagger.v3.oas.annotations.media.Schema;
 
+@ArraySchema(arraySchema = @Schema(description = "BPN Mappings"), maxItems = Integer.MAX_VALUE)
 public record BpnEdcMapping(@ApiModelProperty(example = "BPNL00000003CSGV") String bpn,
                             @ApiModelProperty(example = "https://trace-x-test-edc.dev.demo.catena-x.net/a1") String url) {