+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Bumped veracode/veracode-uploadandscan-action@0.2.5 to 0.2.6
 - Updated catena links from readme to reflext tractusx links
 - Updated path of digital twin registry for dev environment
+- Updated swagger api doc and added collection
 
 ## [3.3.0] - 2023-05-02
 ### Added